WIP #18

+++ src/main/java/targetImage/CompareResult.java
@@ -9,17 +9,17 @@ import org.opencv.core.Mat;
 import org.opencv.core.Rect;
 
 public class CompareResult {
-	private Map<Rect, Map<String, Float>> resultMaps;
+	private Map<Rect, Map<String, Double>> resultMaps;
 
 	private Mat resultMat;
 
-	public CompareResult(Map<Rect, Map<String, Float>> resultMaps, Mat resultMat) {
+	public CompareResult(Map<Rect, Map<String, Double>> resultMaps, Mat resultMat) {
 		super();
 		this.resultMaps = resultMaps;
 		this.resultMat = resultMat;
 	}
 
-	public Map<Rect, Map<String, Float>> getResultMaps() {
+	public Map<Rect, Map<String, Double>> getResultMaps() {
 		return resultMaps;
 	}
 
@@ -27,12 +27,12 @@ public class CompareResult {
 		return resultMat;
 	}
 
-	public Float getAverage(Rect key) {
+	public Double getAverage(Rect key) {
 		return this.resultMaps.get(key).get("average");
 	}
 
-	public List<Map<String, Float>> getResultList() {
-		List<Map<String, Float>> resultList = new ArrayList<>();
+	public List<Map<String, Double>> getResultList() {
+		List<Map<String, Double>> resultList = new ArrayList<>();
 		Set<Rect> faceKeys = this.resultMaps.keySet();
 
 		for (Rect faceKey : faceKeys) {
